Rename utils.js to utils.jsx. Move highlightSubstring to utils.jsx. Highlight all matching substrings in accordion title & item.
