+++ src/components/autocomplete-item.jsx
@@ -1,23 +1,8 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { getLastIndexOfSubstringIgnoreCase } from '../utils';
+import { highlightSubstring } from '../utils';
 
 class AutocompleteItem extends Component {
-  static highlightSubstring(string, substring) {
-    const i = getLastIndexOfSubstringIgnoreCase(string, substring);
-    if (i < 0) return string;
-    const prestring = string.slice(0, i);
-    const highlight = string.slice(i, i + substring.length);
-    const poststring = string.slice(i + substring.length);
-    return (
-      <Fragment>
-        {prestring}
-        <b>{highlight}</b>
-        {poststring}
-      </Fragment>
-    );
-  }
-
   componentDidMount() {
     this.ensureVisible();
   }
@@ -49,7 +34,7 @@ class AutocompleteItem extends Component {
           className={active ? 'hover' : ''}
         >
           {substringToHighlight
-            ? AutocompleteItem.highlightSubstring(item.pathLabel, substringToHighlight)
+            ? highlightSubstring(item.pathLabel, substringToHighlight)
             : item.label}
         </button>
       </li>
